+++ node/ts/net.ts
@@ -409,7 +409,24 @@ export class Net {
     Native.ConnectionManager_clear_proxy(this._connectionManager);
   }
 
-  /** Updates the remote config settings used by libsignal. */
+  /**
+   * Updates libsignal's remote configuration settings.
+   *
+   * The provided configuration map must conform to the following requirements:
+   * - Each key represents an enabled configuration and directly indicates that the setting is enabled.
+   * - Keys must have had the platform-specific prefix (e.g., `"desktop.libsignal."`) removed.
+   * - Entries explicitly disabled by the server must not appear in the map.
+   * - Values originally set to `null` by the server must be represented as empty strings.
+   * - Values should otherwise maintain the same format as they are returned by the server.
+   *
+   * These constraints ensure configurations passed to libsignal precisely reflect enabled
+   * server-provided settings without ambiguity.
+   *
+   * Only new connections made *after* this call will use the new remote config settings.
+   * Existing connections are not affected.
+   *
+   * @param remoteConfig A map containing preprocessed libsignal configuration keys and their associated values.
+   */
   setRemoteConfig(remoteConfig: Map<string, string>): void {
     Native.ConnectionManager_set_remote_config(
       this._connectionManager,